Suggested update
I have now been through this properly. I still prefer sorting the children once
when we find them. The order we get the children nodes back from the LSP is
random, so there is no reason to prefer the LSP's sorting to our own sorting.
Additionally the sorting algorithm was being called recursively in every level
expansion that is being rendered and this just seems wasteful. We even sort
duplicate nodes (that were both from the same cache location).

I do like the idea of encoding the node to get the filename sorting included. I
dropped the need to pass the filename into the quicksort function by making the
function a class method, which then knows what its root is. I also dropped the
filename part where we are in the same file as the root, since this will be
guaranteed to be smaller than any other file that has the same 5 digit line
number _plus_ a file name.

On proper reflection, I decided that we did not need to sort children when
adding from a cache entry. This is for the simple reason that the cache entry
itself will already be sorted. This means we only ever need to sort children the
very first time we search a new node and ever after its children will remain
sorted.

Please review and test against your own set up to make sure I am not introducing
any bugs or regressions. I also welcome your criticism if you think I am making
the wrong strategic decision. I envisage that you would squash (or fixup) the 4
commits of this chain before we merge in to main, to keep the history a bit
cleaner